requirements.txt
# requirements.txt

# Библиотека для взаимодействия с Telegram API
aiogram==3.0.0b7

# Библиотека для загрузки переменных окружения из .env файла
python-dotenv==1.0.0

# Библиотека для взаимодействия с панелью 3x-ui
py3xui==0.1.0

# Необходимые зависимости для py3xui (если py3xui не включает их автоматически)
aiohttp==3.8.4

# Дополнительные зависимости (опционально, если используются)
# logging, json, asyncio, uuid, dataclasses — встроенные модули Python, не требуют установки

